+++ API/API.lua
@@ -4,15 +4,19 @@
 ---@alias betterCommandDef {description: string, param?: string, privs: table<string, boolean>, func: betterCommandFunc}
 
 --local bc = better_commands
-local storage = minetest.get_mod_storage()
 
 
 local modpath = minetest.get_modpath("better_commands")
+
+---Runs a file
+---@param file string
+---@param subfolder string?
 function better_commands.run_file(file, subfolder)
     dofile(string.format("%s%s%s.lua", modpath, subfolder and "/"..subfolder.."/" or "", file))
 end
 
 local api_files = {
+    "storage",
     "damage",
     "entity",
     "parsing",
@@ -20,24 +24,15 @@ local api_files = {
     "scoreboard",
     "teams",
     "scoreboard_criteria",
+    "spawnpoint",
 }
 
-for _, file in ipairs(api_files) do
-    better_commands.run_file(file, "API")
-end
+better_commands.registered_on_update = {}
 
-local scoreboard_string = storage:get_string("scoreboard")
-if scoreboard_string and scoreboard_string ~= "" then
-    better_commands.scoreboard = minetest.deserialize(scoreboard_string)
-else
-    better_commands.scoreboard = {objectives = {}, displays = {colors = {}}}
-end
-
-local team_string = storage:get_string("teams")
-if team_string and team_string ~= "" then
-    better_commands.teams = minetest.deserialize(team_string)
-else
-    better_commands.teams = {teams = {}, players = {}}
+---Registers a function to be called every save_interval
+---@param func function
+function better_commands.register_on_update(func)
+    table.insert(better_commands.registered_on_update, func)
 end
 
 local timer = 0
@@ -45,18 +40,16 @@ minetest.register_globalstep(function(dtime)
     timer = timer + dtime
     if timer > better_commands.settings.save_interval then
         timer = 0
-        storage:set_string("scoreboard", minetest.serialize(better_commands.scoreboard))
-        storage:set_string("teams", minetest.serialize(better_commands.teams))
-        better_commands.update_hud()
+        for _, func in ipairs(better_commands.registered_on_update) do
+            func()
+        end
     end
-end)
-
-minetest.register_on_shutdown(function()
-    storage:set_string("scoreboard", minetest.serialize(better_commands.scoreboard))
-    storage:set_string("teams", minetest.serialize(better_commands.teams))
-    storage:set_string("successful_shutdown", "true")
 end)
 
 minetest.register_on_joinplayer(function(player)
     better_commands.sidebars[player:get_player_name()] = {}
 end)
+
+for _, file in ipairs(api_files) do
+    better_commands.run_file(file, "API")
+end