+++ API/spawnpoint.lua
@@ -1,0 +1,13 @@
+minetest.register_on_mods_loaded(function()
+    minetest.register_on_respawnplayer(function(player)
+        local name = player:get_player_name()
+        if better_commands.spawnpoints[name] then
+            player:set_pos(better_commands.spawnpoints[name])
+            return true
+        end
+    end)
+end)
+
+function better_commands.set_spawn(player, pos)
+    better_commands.spawnpoints[player:get_player_name()] = pos
+end